auto convert StoragrError to ActionResult and replace raw HttpCode numbers with "StatusCodes.Status...."

+++ src/Storagr.Store/Controllers/ObjectController.cs
@@ -1,6 +1,8 @@
 ﻿using System.Collections.Generic;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Storagr.Shared;
 using Storagr.Shared.Data;
 using Storagr.Store.Services;
 
@@ -19,34 +21,38 @@ namespace Storagr.Store.Controllers
         }
 
         [HttpGet]
-        [ProducesResponseType(200, Type = typeof(IEnumerable<StoreObject>))]
-        [ProducesResponseType(404)]
+        [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(IEnumerable<StoreObject>))]
+        [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(StoragrError))]
         public IActionResult List([FromRoute] string repositoryId)
         {
             if (!_storeService.Exists(repositoryId))
-                return NotFound();
+                return (ActionResult) new RepositoryNotFoundError();
 
             return Ok(_storeService.List(repositoryId));
         }
 
         [HttpGet("{objectId}")]
-        [ProducesResponseType(200, Type = typeof(StoreObject))]
-        [ProducesResponseType(404)]
+        [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(StoreObject))]
+        [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(StoragrError))]
         public IActionResult Get([FromRoute] string repositoryId, [FromRoute] string objectId)
         {
+            if (!_storeService.Exists(repositoryId))
+                return (ActionResult) new RepositoryNotFoundError();
             if (!_storeService.Exists(repositoryId, objectId))
-                return NotFound();
+                return (ActionResult) new ObjectNotFoundError();
 
             return Ok(_storeService.Get(repositoryId, objectId));
         }
         
         [HttpDelete("{objectId}")]
-        [ProducesResponseType(200)]
-        [ProducesResponseType(404)]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(StoragrError))]
         public IActionResult Delete([FromRoute] string repositoryId, [FromRoute] string objectId)
         {
+            if (!_storeService.Exists(repositoryId))
+                return (ActionResult) new RepositoryNotFoundError();
             if (!_storeService.Exists(repositoryId, objectId))
-                return NotFound();
+                return (ActionResult) new ObjectNotFoundError();
 
             _storeService.Delete(repositoryId, objectId);
