+++ src/Storagr/Controllers/ObjectController.cs
@@ -2,8 +2,10 @@
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Storagr.Services;
+using Storagr.Shared;
 using Storagr.Shared.Data;
 
 namespace Storagr.Controllers
@@ -24,13 +26,13 @@ namespace Storagr.Controllers
         
         
         [HttpGet]
-        [ProducesResponseType(200, Type = typeof(StoragrObjectListResponse))]
-        [ProducesResponseType(404, Type = typeof(StoragrError))]
+        [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(StoragrObjectListResponse))]
+        [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(StoragrError))]
         public async Task<IActionResult> List([FromRoute] string repositoryId, [FromQuery] StoragrObjectListRequest request)
         {
             var repository = await _objectService.Get(repositoryId);
             if (repository == null)
-                return NotFound(new RepositoryNotFoundError());
+                return (ActionResult) new RepositoryNotFoundError();
             
             var user = await _userService.GetAuthenticatedUser();
             var objects = (await _objectService.GetAll(repositoryId));
@@ -57,42 +59,50 @@ namespace Storagr.Controllers
         }
 
         [HttpGet("{objectId}")]
-        [ProducesResponseType(200, Type = typeof(StoragrObject))]
-        [ProducesResponseType(404)]
+        [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(StoragrObject))]
+        [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(StoragrError))]
         public async Task<IActionResult> Get([FromRoute] string repositoryId, [FromRoute] string objectId)
         {
             var obj = await _objectService.Get(repositoryId, objectId);
             if (obj == null)
-                return NotFound();
+                return (ActionResult) new ObjectNotFoundError();
 
             return Ok((StoragrObject)obj);
         }
         
         [HttpPost("{objectId}")]
-        [ProducesResponseType(200)]
-        [ProducesResponseType(400)]
-        [ProducesResponseType(404)]
-        public async Task<IActionResult> Verify([FromRoute] string repositoryId, [FromRoute] string objectId, [FromBody] StoragrObjectVerifyRequest verifyRequest)
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(StoragrError))]
+        [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(StoragrError))]
+        public async Task<IActionResult> Verify([FromRoute] string repositoryId, [FromRoute] string objectId, [FromBody] StoragrObject expectedObject)
         {
-            if (objectId != verifyRequest.ObjectId)
+            if (objectId != expectedObject.ObjectId)
                 return BadRequest();
 
             var repository = await _objectService.Get(repositoryId);
             if (repository == null)
-                return NotFound(new RepositoryNotFoundError());
+                return (ActionResult) new RepositoryNotFoundError();
 
-            if (await _objectService.Create(repository.RepositoryId, verifyRequest.ObjectId, verifyRequest.Size) == null)
-                return NotFound(new ObjectNotFoundError());
+            if (await _objectService.Create(repository.RepositoryId, expectedObject.ObjectId, expectedObject.Size) == null)
+                return (ActionResult) new ObjectNotFoundError();
 
             return Ok();
         }
 
         [HttpDelete("{objectId}")]
         [Authorize(Policy = "Management")]
-        [ProducesResponseType(204)]
-        [ProducesResponseType(500)]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(StoragrError))]
         public async Task<IActionResult> Delete([FromRoute] string repositoryId, [FromRoute] string objectId)
         {
+            var repository = await _objectService.Get(repositoryId);
+            if (repository == null)
+                return (ActionResult) new RepositoryNotFoundError();
+            
+            var obj = await _objectService.Get(repositoryId, objectId);
+            if (obj == null)
+                return (ActionResult) new ObjectNotFoundError();
+            
             try
             {
                 await _objectService.Delete(repositoryId, objectId);
